pyproject.toml
[tool.poetry]
name = "mirage"
version = "0.1.0"
description = ""
authors = ["Tarun Luthra <tarunluthra987@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
black = "^21.7b0"
django-rest-framework = "^0.1.0"
python-dotenv = "^0.19.0"
django-cors-headers = "^3.8.0"
psycopg2-binary = "^2.9.1"
requests = "^2.26.0"
uuid = "^1.30"
django-audit-log = "^0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
